invert server and client directions so packets always flow from client to server
